# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OCMS — Construction ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/dprs:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [DPR]
      operationId: listProjectDprs
      responses:
        '200':
          description: DPRs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dpr' }
    post:
      tags: [DPR]
      operationId: createDpr
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DprInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dpr' }
  /dprs/{dprId}:
    parameters:
      - name: dprId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [DPR]
      operationId: getDpr
      responses:
        '200':
          description: DPR detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DprDetail' }
  /dprs/{dprId}/submit:
    parameters:
      - name: dprId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [DPR]
      operationId: submitDpr
      responses:
        '200':
          description: Submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dpr' }
  /dprs/{dprId}/approve:
    parameters:
      - name: dprId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [DPR]
      operationId: approveDpr
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DprApprovalInput' }
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dpr' }
...
  /projects/{projectId}/documents:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Documents]
      operationId: listProjectDocuments
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
    post:
      tags: [Documents]
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
  /projects/{projectId}/documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
      - name: documentId
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Documents]
      operationId: deleteDocument
      responses:
        '204':
          description: Deleted
...
  /approvals/{approvalId}/resolve:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Approvals]
      operationId: resolveApproval
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalResolutionInput' }
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalItem' }
...
components:
  parameters:
...
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id: { type: string }
        email: { type: [string, 'null'], format: email }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        profileImageUrl: { type: [string, 'null'] }
...
    UserRole:
      type: string
      enum: [owner, pm, site_engineer, qs, finance, contractor, qc, store, hr, admin]
...
    DprStatus:
      type: string
      enum: [draft, submitted, approved, rejected]
...
    Dpr:
      type: object
      required: [id, projectId, reportDate, weather, temperature, manpowerCount, summary, status, submittedById, approvedById, submittedAt, approvedAt, rejectionReason, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        reportDate: { type: string, format: date-time }
        weather: { type: [string, 'null'] }
        temperature: { type: [number, 'null'] }
        manpowerCount: { type: integer }
        summary: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/DprStatus' }
        submittedById: { type: [string, 'null'] }
        approvedById: { type: [string, 'null'] }
        submittedAt: { type: [string, 'null'], format: date-time }
        approvedAt: { type: [string, 'null'], format: date-time }
        rejectionReason: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    DprInput:
      type: object
      required: [reportDate]
      properties:
        reportDate: { type: string, format: date-time }
        weather: { type: string }
        temperature: { type: number }
        manpowerCount: { type: integer }
        summary: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/DprItemInput' }
    DprApprovalInput:
      type: object
      properties:
        approve: { type: boolean }
        rejectionReason: { type: string }
    DprItemInput:
      type: object
      required: [activityId, quantityToday]
      properties:
        activityId: { type: string }
        quantityToday: { type: number }
        remarks: { type: string }
    DprItem:
      type: object
      required: [id, dprId, activityId, activityName, quantityToday, cumulativeQuantity, remarks]
      properties:
        id: { type: string }
        dprId: { type: string }
        activityId: { type: string }
        activityName: { type: string }
        quantityToday: { type: number }
        cumulativeQuantity: { type: number }
        remarks: { type: [string, 'null'] }
    DprDetail:
      type: object
      required: [dpr, items, photos, submittedBy, approvedBy]
      properties:
        dpr: { $ref: '#/components/schemas/Dpr' }
        items:
          type: array
          items: { $ref: '#/components/schemas/DprItem' }
        photos:
          type: array
          items: { $ref: '#/components/schemas/SitePhoto' }
        submittedBy:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
        approvedBy:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    SitePhoto:
      type: object
      required: [id, projectId, activityId, dprId, url, caption, capturedAt, latitude, longitude, uploadedById, tag, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        activityId: { type: [string, 'null'] }
        dprId: { type: [string, 'null'] }
        url: { type: string }
        caption: { type: [string, 'null'] }
        capturedAt: { type: string, format: date-time }
        latitude: { type: [number, 'null'] }
        longitude: { type: [number, 'null'] }
        uploadedById: { type: [string, 'null'] }
        tag: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
...
    Document:
      type: object
      required: [id, projectId, name, category, url, version, uploadedById, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        category: { type: [string, 'null'] }
        url: { type: string }
        version: { type: integer }
        uploadedById: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    DocumentInput:
      type: object
      required: [name, url]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        url: { type: string, minLength: 1 }
        version: { type: integer }
...
    ApprovalItem:
      type: object
      required: [id, projectId, projectName, entityType, entityId, title, requestedById, requestedByName, assignedToRole, status, createdAt, resolvedAt, ageDays]
      properties:
        id: { type: string }
        projectId: { type: [string, 'null'] }
        projectName: { type: [string, 'null'] }
        entityType: { type: string }
        entityId: { type: string }
        title: { type: string }
        requestedById: { type: [string, 'null'] }
        requestedByName: { type: [string, 'null'] }
        assignedToRole: { $ref: '#/components/schemas/UserRole' }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        ageDays: { type: integer }
    ApprovalResolutionInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, rejected] }
        comment: { type: string }
...